chore: #94 move block_signature to separate file within block dir
