+++ src/utils/mod.rs
@@ -1,2 +1,1 @@
 pub mod hash;
-pub mod signature;
